refactor(monitoring): standardize repository access
